cloud-function/requirements.txt
Flask==2.0.2
functions-framework==3.0.0
pyheif==0.6.1
Pillow==8.4.0
Werkzeug==2.0.2